requirements.txt
Django>=5.1.1
djangorestframework>=3.15.2
numpy>=2.1.1
